eilmer: Pass user "dt" value for user-defined grid motion.
The user also gets access to the current dt_global if
required when using the user-defined function,
assignVtxVelocities(), in the Lua interface.

This might help in cases where the user knows something
about the displacement of vertices over the timestep
but wishes to convert to a velocity. Now the option to
divide by dt is available. This is because the interface
formulation is based on the user-supplying vertex
velocities.