geom: Added to error messages in Line constructor.
As part of the mission to add a better error messages as
we come across them, I added some detail to the Line constructor.
I was using the Line constructor but I had a 'nil' value passed
to 'p1'. The error message was that 'p1' was missing, but when I
looked at my script, I could definitely see a 'p1'. If you assign
a nil to a key in at Lua table, lua will wipe that key out. So by
the time the constructor got to look for 'p1' it was indeed missing.
The error message now tells the user to check also that the value
passed to 'p1' is valid.