Eilmer: In the Lua domain, introduce FluidBlock to replace SBlock and UBlock.
This gives us a more regular set of names for use in the input scripts.

Old name          New Name
-----------------------------------
SBlock            FluidBlock
UBlock            FluidBlock
SBlockArray       FluidBlockArray
SSolidBlock       SolidBlock
SSolidBlockArray  SolidBlockArray
-----------------------------------

For a short while, the old names are retained as aliases so that all current
scripts should continue to work, for now.
The arguments to the functions are unchanged.