Geom: Can now import individual StructuredGrid objects in your Lua script.
For example:
importedGrid = StructuredGrid:new{file="myFileName.dat", fmt="vtk"}

The format "vtk" is the legacy text format for the VTK system.
Other format values could be "gzip", which is the Eilmer4 native file
or "text", which is essentially the Eilmer3 native file uncompressed.

There is an example of running the cone20 case with grids that were
generated for Eilmer3.
See examples/eilmer/2D/sharp-cone-20-degrees/sg-import-grid/