eilmer: Moved compute_leastsq_weights
compute_leastsq_weights() is used to calculate the weights for the
least-squares calculation of flow gradients for the viscous terms. Initially
this was performed at the end of compute_primary_cell_geometric_data(), however
at this stage the ghost cells do not have the correct cell centre positions.
Which meant that the weights were incorrect (by a small amount) at the edge
of blocks.

This update moves compute_leastsq_weights() to the top of
compute_least_squares_setup_for_reconstruction(), since at this point the
cells (including ghost cells) have the correct cell positions.