Eilmer: Set FromStagnationBC to give flow into the domain.
Ingo noted that issue of normals pointing into the domain and
pointing out of the domain on different boundaries meant
that EAST boundaries weren't playing nicely with the subsonic
inflow BC. The fix is to reverse the flux direction based on
whether a boundary has an "into-the-domain" or "out-of-the-domain"
normal. I have called the latter an outwardNormal in the implementation.