Eilmer: Added user-configurable control on diffuse wall-bcs initialisation.
A user can request that the wall boundary conditions are diffused
into the domain upon initialisation in the simulation runtime.
Note we do this at simulation startup and not during the
"--prep" stage because we need access to all of the cells,
interfaces and their connections to one another.

To request this feature, set:

   config.diffuse_wall_bcs_on_init = true

The diffusion of wall BC occurs in a number of passes. On each pass,
a new layer of cells away from the wall is included in the flow
averaging proces. The default number of passes is set at 30 because this
is usually a good number of cells to include in a laminar
boundary layer. The user can control this number of passes with:

   config.number_init_passes = 20