Eilmer: Turbulence properties only need special care on step == 0
In Rev. 1049, I fixed a bug such that the turbulence properties
were available before the convective update so that they could
be averaged and placed in the interfaces. This only needs to
occur on the very first step. I had that occurring on every step.

This didn't cause any problems with accuracy, but it did add
redundant calculations to the simulation.