eilmer: Refactored the code for assembling the usg viscous derivative clouds.
I have refactored the code for assembling the usg viscous derivatives clouds,
used in the weight-least squares method.

This commit also adds a check to ensure clouds on the domain boundary have more
than three points. If there is a boundary interface that has three or less
points in it's cloud, then the rules for adding points to the cloud that ensure
2nd order spatial accuracy are relaxed, and the cloud is grown.