Eilmer/Gas: Added GasException, and catch this in decode_conserved.
I wanted to add a bit more detail to some points in the simulation
code where an update fails. Prior to this commit, the first time
we knew that temperature had gone bad was when we'd try to evaluate
viscosity with a negative temperature value. We shouldn't be handing
bad values like that to the viscosity routines.

Now there is a GasException class that can be used to signal
problems in the gas modelling code. In this commit, I've added
an exception when the ideal gas model computes a negative
temperature and/or pressure.