Eilmer: Added more robustness to the steady-state solver.
There are a few points where the steady-state solver might
try to step to large in time, or make a perturbation to the
conserved variables that is too large. In these cases,
cell.decode_conserved will throw an exception. These exceptions
are now all caught and a decision is made. The decision usually
involves reducing the size of the change. If it is fails repeatedly
after a number of attempts, then the code will exit and print
some message.