eilmer: CUDA gpu-chem module now executes 1D array of blocks
Initially I had the 1D array of kernels being executed as one block,
however the max number of kernels for a block is 1024. This push updates
the 1D arrays to be stored as a grid (referring to a CUDA block/grid, not a
CFD block/grid!), whereby each kernel instantiation is a block rather than a
thread. This allows us to run simulations with more than 1024 cells.