Gas: Very-viscous-air can now be made up of multiple components.
We'd like to use the very-viscous-air model in a manufactured
solution to test the mass diffusion and catalytic boundary conditions.
For these tests, we wanted multiple species. So now the very-viscous-air
model can be made up of multitple identical components. The different
components may be thought of as tagging different parts of the gas
flow.