Eilmer: Solid domains now handle anisotropic material properties.
The infrastructure change to accommodate spatially-varying material
properties in the solid domains is to move the properties from
a being a block attribute to a cells and interfaces attribute.
Presently, the interface properties are constructed as simple averages
of the attached cells. We might need to think about a length- or
area-weighted average for non-uniform grids.

Also, I should think about converting the SolidProps from a class
to a struct now that there is one of these for every cell and
interface.