gas: Added bare-bones of a vibrationally-specific nitrogen model.
This is the shell of a gas model so that Katrina Sklavos
can work on the implementation for her thesis work. Presently,
we need the thermodynamic behaviour filled in.

In the next step, we'll think about adding the kinetic behaviour,
that is, how the specific vibrational levels relax against the
translational mode and against each other.