Geom: Fix the calculation of vector areas of triangles (factor of 2).
This fixes a false alarm for Jens' nozzle grid.  We had thought that the
pyramid cells were not being correctly imported from the SU2 file, however,
it was the closed-area check that was failing with mixed triangular and
quadrilateral facets.

Along the way, some of the code was refactored and better messages were added
to the thrown exception.