geom: Added an ij rotation and swap faces function for structured grids.
Sometimes when using Gridpro software to generate grids, the i,j axis of
individual blocks may not be consistent across the domain. However, when
doing shock fitting simulations we do require a very specific arrangement
i and j indicies. The rotation and swap faces functions allows the axis
to be reoriented. In this way we can transform 2D structured grids to have
the correct i,j orientations.