Eilmer: introduce reverse look-up from vertices to faces and cells.
Actually, we moved the reverse look-up from vertices to faces from the
UnstructuredGrid class into the flow solver, then added the look-up from
vertices to cells for both unstructured and structured grids.

Kyle had been using the vertex-to-face look-up for building clouds of points
for gradient calculations and that functionality should be unchanged.