gas: Added error checks for valid entries in GasState table for Lua wrappers.
I added a number of error checks to the user-supplied data
before we call through to the D methods when we can reasonably
expect valid data. For example, using a pressure value that is negative,
zero or completely unset (initialised as NaN) is NOT a good
starting point for calling a thermo update that uses pressure
as an input.