adjoint: updated adjoint solver for 2D mesh.
Updated the adjoint solver to handle a 2D mesh. The comparison with
finite-difference gradients for the nozzle problem on a 100x8 mesh
yields ~ 0.1% error, the SNU lab members tell me that when using FD verification
the solution should be mesh indepedent, currently the adjoint solver
is too inefficient to run a proper test, but when comparing meshes of size
100x1, 100x4, 100x8 the error between adjoint and FD gradients is ~0.25%,
~0.15%, ~0.1% respectively. This gives me some confidence that, as the code
stands, the adjoint solver is performing correctly for single-block, 2D,
Euler problems.