adjoint: Improved efficiency of flux sensitivity to a vertex perturbation
Previously when perturbing a vertex, we would loop through the interfaces
that shared that vertex to compute the new interface flux, during that loop
we updated the geometry of all interfaces and cells attached to that vertex.
The code now only updates the geometry of the current interface, and it's left
and right cell neighbours.