Eilmer: Add limiting on mu_effective values in k-omega model.
The mu_effective values are used in the calculation of
the diffusive flux transport of the turbulence properties
tke and omega. These mu_effective values are subtly different
from the turbulent viscosity value. We had applied limiting
to the turbulent viscosity value, but no such limiting on
these mu_effective values. In this commit, we now perform
that limiting on the mu_effective values also, based on the
max_mu_t_factor.