nm: added in ILU(p) factorisation
ILU(p) factorisation allows for the user to specify the level of
fill permitted. As p increases the computational time increases,
but so to does the accuracy of the factorisation. If p = n-1, where n
is the number of columns or rows in the matrix to be factorised, then
this method will compute the exact factorisation.

NB: currently there is some inefficiency as we construct the factorisation
using a full matrix format, and then convert this to sparse matrix format.