Eilmer: Default is to compile with 'fast'; dropped e4sss and e4adjoint in default build
Since we are close to a production release (4.0), we have set the default compile
target to 'fast'. There is a note in the INSTALL file describing why and how one
would do a 'debug' build.

Also, because the the 'fast' build actually takes a bit of time, we dropped
the e4sss and e4adjoint programs from the list of default-build programs.
To request these programs explicitly, use the make variable WITH_SSS and
WITH_ADJOINT. For example:

  $ make WITH_SSS=1 WITH_ADJOINT=1 install