eilmer: only store the transposed Jacobian in adjoint solver
To get the ball rolling the adjoint solver was initially coded naively, in a
memory hungry fashion -- we were storing 3 versions of the
Jacobian (one full Jacobian matrix, one full transposed Jacobian matrix,
and one sparse Jacobian matrix). We now only store the sparse transposed
Jacobian. We are able to store the transposed Jacobian directly by
computing and storing the Jacobian in Compressed Column Storage format,
this is the equivalent of storing the transposed Jacobian in Compressed
Row Storage format, fortuitously CSR is the format we used in our sparse linear
algebra methods.