Gasflow: weed out a troublesome limiting case when using the CEA gas model.
When using the CEA gas model, we can get numbers that are slightly different
depending on the way we ask the calculations to be done.  The limiting case
that was giving trouble was the expansion from stagnation to a specified
pressure, where the new pressure was nominally the same as the stagnation
pressure.  This new condition had static enthalpy just a little larger than
the stagnation enthalpy and, when computing the velocity, we were taking the
square-root of a negative number and getting a nan.  Once we had a nan, it
propagated and any conditional tests were just evaluating to false, giving us
badly behaved numerical methods that just ploughed on with rubbish data.

I suppose that the lesson is to handle limiting cases carefully.

Also, along the way, the bracketing function template was cleaned up a little.
It now swaps the values to be consistent with the expected ordering.