eilmer: added arrays for storing local interface clouds for adjoint solver
This is the first stage (out of two) for improving the efficiency of the adjoint solver by
attaching an array of effected interfaces to fvcell and fvvertex objects. To compute the
perturbed flux for forming the Jacobians, we had been looping
around the ifaces of cells in the jacobian_stencil array, however this leads to many redundant
computations. We now store an array of effected interfaces, in addition to effected cells. This
will remove the redundant computations.

This first stage adds in the code for collecting the interfaces into the arrays attached to the
fvcell and fvvertex objects.